% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xomultinom_classes.R
\name{as.data.frame.xomultinom_dist}
\alias{as.data.frame.xomultinom_dist}
\title{Coerce an \code{xomultinom_dist} object to a data frame}
\usage{
\method{as.data.frame}{xomultinom_dist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{xomultinom_dist}.}

\item{...}{Further arguments passed to or from other methods (currently
unused).}
}
\value{
A \code{data.frame} with columns \code{x} (evaluation points) and
either \code{pmf} or \code{cdf} (probability values). If the object was
computed on the log scale the column is named \code{log_pmf} or
\code{log_cdf} accordingly.
}
\description{
Converts the evaluation points and probability values stored in an
\code{xomultinom_dist} object into a tidy \code{data.frame} suitable for
further manipulation or export.
}
\examples{
k <- 5; n <- 40
obj <- maxmultinomcdf(size = n, prob = rep(1/k, k))
head(as.data.frame(obj))

}
