% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_decr.R
\name{decr_2_pmin}
\alias{decr_2_pmin}
\title{Utility function.}
\usage{
decr_2_pmin(m, decr)
}
\arguments{
\item{m}{An integer vector of numbers of multinomial classes.}

\item{decr}{A numeric vector of decrement values for which to compute
the probabilities.}
}
\value{
A numeric matrix.
}
\description{
This is an auxiliary function to compute the probability values from
an decrement with respect to the case of equiprobability.
}
\examples{
m <- 3:50
decr <- seq(0, 1, 0.01)
pmin <- decr_2_pmin(m, decr)
summary(as.numeric(pmin))

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \doi{10.1098/rsos.190198}.
}
\seealso{
\code{\link{pmin_2_decr}} for the opposite calculation.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
