% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dminmultinom_corrado}
\alias{dminmultinom_corrado}
\title{Minimum probability for multinomial distribution}
\usage{
dminmultinom_corrado(x, size, prob, logd, verbose)
}
\arguments{
\item{x}{Numeric vector of values at which to evaluate the probability.}

\item{size}{Integer, the total number of trials.}

\item{prob}{Numeric vector of category probabilities.}

\item{logd}{Logical, if TRUE returns log-probabilities.}

\item{verbose}{Logical, if TRUE shows a progress bar.}
}
\value{
Numeric vector of probabilities (or log-probabilities if \code{logd = TRUE}).
}
\description{
Computes the probability that the minimum cell count in a multinomial
sample is equal to a given value.
}
