% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{drangemultinom_corrado}
\alias{drangemultinom_corrado}
\title{Range probability for multinomial distribution}
\usage{
drangemultinom_corrado(x, size, prob, logd, verbose)
}
\arguments{
\item{x}{Numeric vector of range values at which to evaluate the probability.}

\item{size}{Integer, the total number of trials.}

\item{prob}{Numeric vector of category probabilities.}

\item{logd}{Logical, if TRUE returns log-probabilities.}

\item{verbose}{Logical, if TRUE shows a progress bar.}
}
\value{
Numeric vector of probabilities (or log-probabilities if \code{logd = TRUE}).
}
\description{
Computes the probability that the range of a multinomial sample
(i.e., max(X) - min(X)) is equal to a given value.
}
