% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{highest_order_statistics}
\alias{highest_order_statistics}
\title{CDF of the sum of the first \eqn{J} largest order statistics}
\usage{
highest_order_statistics(td, n, m, J)
}
\arguments{
\item{td}{Numeric value at which to evaluate the CDF. Internally, this is
converted to \code{floor(td)}.}

\item{n}{Integer number of balls/trials.}

\item{m}{Integer number of urns/cells.}

\item{J}{Integer number of largest order statistics to sum.}
}
\value{
Numeric value giving the CDF probability.
}
\description{
Computes the cumulative distribution function of the sum of the first
\eqn{J} largest order statistics for an equiprobable multinomial
distribution.
}
\examples{
highest_order_statistics(6, 10, 5, 2)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019). Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics. Royal Society
Open Science, 6, 190198. \doi{10.1098/rsos.190198}
}
