% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_decr.R
\name{incr_2_pmax}
\alias{incr_2_pmax}
\title{Utility function.}
\usage{
incr_2_pmax(m, incr)
}
\arguments{
\item{m}{An integer vector of numbers of multinomial classes.}

\item{incr}{A numeric vector of increment values for which to compute
the probabilities.}
}
\value{
A numeric matrix.
}
\description{
This is an auxiliary function to compute the probability values from
an increment with respect to the case of equiprobability.
}
\examples{
m <- 3:50
incr <- seq(0, 1, 0.01)
pmax <- incr_2_pmax(m, incr)
summary(as.numeric(pmax))

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \doi{10.1098/rsos.190198}.
}
\seealso{
\code{\link{pmax_2_incr}} for the opposite calculation.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
