% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{leukaemia}
\alias{leukaemia}
\title{Data: Leukaemia cases}
\format{
{
A data frame with 790 observations and the following 5
variables:
\itemize{
\item \strong{\code{ID}} (\code{int}): 10 character long identification number for a cell
or census district in the study area
\item \strong{\code{x}} (\code{num}): x-coordinate of the geographic centroid of each cell
\item \strong{\code{y}} (\code{num}): y-coordinate of the geographic centroid of each cell
\item \strong{\code{pop}} (\code{int}): 1980 U.S. Census population count for each cell
\item \strong{\code{cases}} (\code{num}): incident cases of leukemia (all types) occurring
between 1978 and 1982 in each cell; fractional values can occur due to
partially missing data
}
}
}
\source{
The data set has been downloaded from \url{https://www.stats.ox.ac.uk/pub/datasets/csb/}.
}
\usage{
data(leukaemia)
}
\description{
This is a well-known epidemiological dataset of diagnosed leukaemia cases
over eight counties in upstate New York. These data originated from the
New York State Cancer Registry, and were gathered during the 5-year period
1978–1982, with a total of 584 individuals diagnosed with leukaemia over a
population of approximately 1 million people. The original data contain
spatial information about registered events split into 790 census tracts.
}
\references{
Lange, N., Ryan, L., Billard, L., Brillinger, D., Conquest, L.,
Greenhouse, J. (1994), "Case Studies in Biometry",
Hoboken, NJ: Wiley & Sons.
}
\keyword{datasets}
