% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin_sample_size.R
\name{max_count}
\alias{max_count}
\title{Compute the Largest Bin Count}
\usage{
max_count(brks, samp_scores, m)
}
\arguments{
\item{brks}{A numeric vector of break points defining the intervals.}

\item{samp_scores}{A numeric vector of sample scores to be assigned to bins.}

\item{m}{An integer specifying the expected number of bins.}
}
\value{
A single integer giving the maximum number of observations contained
in any bin.
}
\description{
Assigns a sample of scores to intervals defined by a set of break points and
returns the size of the largest resulting bin.
}
\details{
The function uses \code{cut()} to classify observations into bins and
\code{tabulate()} to count the number of observations in each bin.
}
