% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_order_statistic}
\alias{max_order_statistic}
\title{CDF of the maximum multinomial cell count}
\usage{
max_order_statistic(td, n, m)
}
\arguments{
\item{td}{Numeric value at which to evaluate the CDF. Internally, this is
converted to \code{floor(td)}.}

\item{n}{Integer number of balls/trials.}

\item{m}{Integer number of urns/cells.}
}
\value{
Numeric value giving \eqn{P(N_{(m)} \le t)}, where
\eqn{t = \lfloor td \rfloor}.
}
\description{
Computes the cumulative distribution function of the largest order
statistic, that is, the maximum cell count, for an equiprobable
multinomial distribution.
}
\examples{
max_order_statistic(3, 10, 5) # P(N_(m) <= 3; n = 10, m = 5)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019). Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics. Royal Society
Open Science, 6, 190198. \doi{10.1098/rsos.190198}
}
