% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pJlargemultinom.R
\name{pJlargemultinom}
\alias{pJlargemultinom}
\title{CDF of the sum of \eqn{J} largest order statistics for a multinomial
distribution}
\usage{
pJlargemultinom(x, size, prob, J = 2, log = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{Numeric vector of values at which to evaluate the CDF.}

\item{size}{Integer number of trials.}

\item{prob}{Numeric vector of non-negative cell probabilities. Values are
internally normalized to sum to 1. Categories with zero probability are
removed before computation.}

\item{J}{Integer number of largest order statistics to consider; default to 2.}

\item{log}{Logical; if \code{TRUE}, returns log-probabilities.}

\item{verbose}{Logical; if \code{TRUE}, displays progress information during
the computation.}
}
\value{
Numeric vector of the same length as \code{x} containing
\eqn{P(S_J \le x)}, \eqn{S_J = \sum_{j=1}^J N_{\langle j\rangle}}, values, or log-probabilities if
\code{log = TRUE}.
}
\description{
Computes the cumulative distribution function of the sum of \eqn{J}
largest order statistics for a multinomial random vector with
equal cell probabilities.
}
\details{
The function works implements the equiprobable case only.
}
\examples{
m <- 4
n <- 60
probs <- rep(1 / m, m)
J <- 3
xseq <- 0:n

cdflarge <- pJlargemultinom(x = xseq, size = n, prob = probs, J = J)
plot(cdflarge)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019). Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics. Royal Society
Open Science, 6, 190198. \doi{10.1098/rsos.190198}
}
\seealso{
\code{\link{pminmultinom}} for the CDF of the minimum,
\code{\link{dmaxmultinom}} for the PMF of the maximum, and
\code{\link{dminmultinom}} for the PMF of the minimum.
}
