% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_decr.R
\name{pmax_2_incr}
\alias{pmax_2_incr}
\title{Utility function.}
\usage{
pmax_2_incr(m, pmax)
}
\arguments{
\item{m}{An integer vector of numbers of multinomial classes.}

\item{pmax}{A numeric vector of probability values for which to compute
the increments.}
}
\value{
A numeric matrix.
}
\description{
This is an auxiliary function to compute the increment as the distance
from the case of equiprobability.
}
\examples{
m <- 3:50
pmax <- seq(0.05, 1, 0.05)
incr <- pmax_2_incr(m, pmax)
summary(as.numeric(incr))

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \doi{10.1098/rsos.190198}.
}
\seealso{
\code{\link{incr_2_pmax}} for the opposite calculation.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
