% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pmin_2_decr}
\alias{pmin_2_decr}
\title{Utility function.}
\usage{
pmin_2_decr(k, pmin)
}
\arguments{
\item{k}{A length-one integer vector of the number of multinomial classes.}

\item{pmin}{A length-one numeric vector with the probability values to
compute the decrements for.}
}
\value{
A numeric matrix.
}
\description{
This is an auxiliary function to compute the decrement as the distance
from the case of equiprobability.
}
\examples{
k <- 3:50
pmin <- seq(0.05, 1, 0.05)
decr <- pmin_2_decr(k, pmin)
summary(as.numeric(decr))

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{decr_2_pmin}} for the opposite calculation.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
