% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob_range_eq}
\alias{prob_range_eq}
\title{Probability mass for the multinomial count range}
\usage{
prob_range_eq(n, pi, r)
}
\arguments{
\item{n}{Integer total count.}

\item{pi}{Numeric vector of category probabilities.}

\item{r}{Numeric count-range value at which to evaluate the probability mass.}
}
\value{
A probability between 0 and 1.
}
\description{
Computes \eqn{P(\max_k N_k - \min_k N_k = r)} for multinomial cell counts
with total count \code{n} and category probabilities \code{pi}.
}
