% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{range_probability}
\alias{range_probability}
\title{CDF of the range of multinomial cell counts}
\usage{
range_probability(td, n, m)
}
\arguments{
\item{td}{Numeric value at which to evaluate the CDF. Internally, this is
converted to \code{floor(td)}.}

\item{n}{Integer number of balls/trials.}

\item{m}{Integer number of urns/cells.}
}
\value{
Numeric value giving \eqn{P(N_{(m)} - N_{(1)} \le t)}, where
\eqn{t = \lfloor td \rfloor}.
}
\description{
Computes the cumulative distribution function of the range, that is,
the difference between the maximum and minimum cell counts, for an
equiprobable multinomial distribution.
}
\examples{
range_probability(1, 7, 10)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019). Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics. Royal Society
Open Science, 6, 190198. \doi{10.1098/rsos.190198}
}
