% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_sum.R
\name{recursive_sum}
\alias{recursive_sum}
\title{Utility function for computing recursively a sum.}
\usage{
recursive_sum(t, n, m, J, sum_depth, cur_depth, rangeArg)
}
\arguments{
\item{t}{A length-one numeric vector indicating the value to compute the
function for.}

\item{n}{A length-one integer vector indicating the number of independent
balls.}

\item{m}{A length-one integer vector indicating the number of independent
urns/cells.}

\item{J}{A length-one integer vector indicating the number of largest order
statistics.}

\item{sum_depth}{A length-one integer vector indicating the summation depth.}

\item{cur_depth}{A length-one integer vector indicating the current
summation depth.}

\item{rangeArg}{Integer vector indicating the range of the summation}
}
\value{
A length-one numeric vector representing the summation value.
}
\description{
This is an auxiliary function to compute a sum recursively.
}
\examples{
highest_order_statistics(6, 10, 5, 2)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{highest_order_statistics}} for computing the CDF
of the sum for the first \eqn{J} largest order statistics for an
equiprobable multinomial distribution.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
