% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_exact}
\alias{round_exact}
\title{Stable rounding function}
\usage{
round_exact(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Integer number of decimal places to round to.}
}
\value{
Numeric vector of rounded values.
}
\description{
Rounds numeric values while avoiding floating-point artifacts by applying
a small offset before rounding.
}
\details{
A small perturbation proportional to the sign of \code{x} is added before
rounding to mitigate issues due to floating-point representation.
}
\examples{
round_exact(0.145, 2)
round_exact(c(1.005, -1.005), 2)

}
