% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{XOMultinom-package}
\alias{XOMultinom-package}
\alias{XOMultinom-pkg}
\alias{XOMultinom}
\title{XOMultinom: Exact distributions of ordered multinomial counts}
\description{
The \pkg{XOMultinom} package provides functions for computing exact
distributions of selected functions of ordered multinomial counts, including
the maximum, minimum, range, and sums of order statistics.

Main functions include:
\itemize{
\item \code{\link{dmaxmultinom}()}, \code{\link{pmaxmultinom}()},
\code{\link{qmaxmultinom}()}, \code{\link{rmaxmultinom}()}
\item \code{\link{dminmultinom}()}, \code{\link{pminmultinom}()},
\code{\link{qminmultinom}()}, \code{\link{rminmultinom}()}
\item \code{\link{drangemultinom}()}, \code{\link{prangemultinom}()},
\code{\link{qrangemultinom}()}, \code{\link{rrangemultinom}()}
\item \code{\link{dJlargemultinom}()}, \code{\link{pJlargemultinom}()},
\code{\link{qJlargemultinom}()}, \code{\link{rJlargemultinom}()}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sergioventurini/XOMultinom}
  \item Report bugs at \url{https://github.com/sergioventurini/XOMultinom/issues}
}

}
\author{
\strong{Maintainer}: Sergio Venturini \email{sergio.venturini@unicatt.it}

Authors:
\itemize{
  \item Sergio Venturini \email{sergio.venturini@unicatt.it}
}

Other contributors:
\itemize{
  \item Marco Bonetti \email{marco.bonetti@unibocconi.it} [contributor]
}

}
\keyword{max}
\keyword{min}
\keyword{multinomial}
\keyword{range}
