% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{decr_2_pmin}
\alias{decr_2_pmin}
\title{Utility function.}
\usage{
decr_2_pmin(k, decr)
}
\arguments{
\item{k}{A length-one integer vector of the number of multinomial classes.}

\item{decr}{A length-one numeric vector with the decrement values to
compute the probabilities for.}
}
\value{
A numeric matrix.
}
\description{
This is an auxiliary function to compute the probability values from
an decrement with respect to the case of equiprobability.
}
\examples{
k <- 3:50
decr <- seq(0, 1, 0.01)
pmin <- decr_2_pmin(k, decr)
summary(as.numeric(pmin))

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{pmin_2_decr}} for the opposite calculation.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
