% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{incr_2_pmax}
\alias{incr_2_pmax}
\title{Utility function.}
\usage{
incr_2_pmax(k, incr)
}
\arguments{
\item{k}{A length-one integer vector of the number of multinomial classes.}

\item{incr}{A length-one numeric vector with the increment values to
compute the probabilities for.}
}
\value{
A numeric matrix.
}
\description{
This is an auxiliary function to compute the probability values from
an increment with respect to the case of equiprobability.
}
\examples{
k <- 3:50
incr <- seq(0, 1, 0.01)
pmax <- incr_2_pmax(k, incr)
summary(as.numeric(pmax))

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{pmax_2_incr}} for the opposite calculation.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
