% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_for_range.R
\name{max_for_range}
\alias{max_for_range}
\title{Utility function for computing the distribution of the range.}
\usage{
max_for_range(t_max, n, m, prev, t)
}
\arguments{
\item{t_max}{A length-one numeric vector indicating the value to compute the
survival function for.}

\item{n}{A length-one integer vector indicating the number of independent
balls.}

\item{m}{A length-one integer vector indicating the number of independent
urns/cells.}

\item{prev}{Integer vector containing the previous iteration values.}

\item{t}{A length-one numeric vector indicating the value to compute for.}
}
\value{
A length-one numeric vector.
}
\description{
This is an auxiliary function to the distribution of the range.
}
\examples{
range_probability(1, 7, 10)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{highest_order_statistics}} for computing the
CDF of the sum of the first \eqn{J} largest order statistics.

\code{\link{smallest_order_value}} for computing the CDF
of the smallest order statistic.

\code{\link{range_probability}} for computing the CDF
of the range.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
