% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_order_statistic.R
\name{max_order_statistic}
\alias{max_order_statistic}
\title{CDF of the highest order statistic for an equiprobable multinomial
distribution.}
\usage{
max_order_statistic(t, n, m)
}
\arguments{
\item{t}{A length-one numeric vector indicating the value to compute the CDF
for.}

\item{n}{A length-one integer vector indicating the number of independent
balls.}

\item{m}{A length-one integer vector indicating the number of independent
urns/cells.}
}
\value{
A length-one numeric vector representing the probability of the
highest order statistic.
}
\description{
This function calculates the cumulative distribution function (CDF) of the
highest order statistic (i.e. the maximum) under an equiprobable multinomial
distribution assumption.
}
\examples{
max_order_statistic(3, 10, 5) # P(N_(1) <= 3; n = 10, m = 5)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{highest_order_statistics}} for computing the
CDF of the sum of the first \eqn{J} largest order statistics.

\code{\link{smallest_order_value}} for computing the CDF
of the smallest order statistic.

\code{\link{range_probability}} for computing the CDF
of the range.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
