% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pmax_2_incr}
\alias{pmax_2_incr}
\title{Utility function.}
\usage{
pmax_2_incr(k, pmax)
}
\arguments{
\item{k}{A length-one integer vector of the number of multinomial classes.}

\item{pmax}{A length-one numeric vector with the probability values to
compute the increments for.}
}
\value{
A numeric matrix.
}
\description{
This is an auxiliary function to compute the increment as the distance
from the case of equiprobability.
}
\examples{
k <- 3:50
pmax <- seq(0.05, 1, 0.05)
incr <- pmax_2_incr(k, pmax)
summary(as.numeric(incr))

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{incr_2_pmax}} for the opposite calculation.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
