% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmaxmultinom.R
\name{pmaxmultinom}
\alias{pmaxmultinom}
\title{Distribution function (CDF) of the maximum for a generic multinomial random
vector.}
\usage{
pmaxmultinom(x, size, prob, log = FALSE, verbose = FALSE, method = "Rcpp")
}
\arguments{
\item{x}{A numeric vector indicating the values of the maximum to compute
the CDF for.}

\item{size}{A length-one integer specifying the total number of objects
that are put into \emph{K} boxes in the typical multinomial experiment.}

\item{prob}{A numeric non-negative of length \emph{K} specifying the
probability for the \emph{K} classes; is internally normalized to sum 1.}

\item{verbose}{A length-one logical vector; if TRUE, details are printed
during the calculation.}

\item{logd}{A length-one logical vector; if TRUE, log probabilities are
computed.}
}
\value{
A numeric vector providing the probabilities of the maximum.
}
\description{
This function calculates the distribution function (i.e. the probability to be
smaller than or equal to a given value) for the maximum of a generic
(i.e. not necessarily equiprobable) multinomial vector.
}
\examples{
k <- 4 # dimension of the multinomial random vector
n <- 60
set.seed(101)
probs <- rdirichlet(1, rep(1/k, k))  # or rep(1/k, k) 
xseq <- 0:n

cdfmax <- pmaxmultinom_C(x = xseq, size = n, prob = probs, log = FALSE,
                       verbose = TRUE)
plot(xseq, cdfmax, type = "s", xlab = "x", ylab = "CDF")

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{pminmulitnom_C}} for computing the
CDF of the minimum for a generic multinomial random vector.

\code{\link{dmaxmulitnom_C}} for computing the
probability mass function (PMF) of the maximum for a generic
multinomial random vector.

\code{\link{dminmulitnom_C}} for computing the
probability mass function (PMF) of the minimum for a generic
multinomial random vector.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
