% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_probability.R
\name{range_probability}
\alias{range_probability}
\title{CDF of the range for an equiprobable multinomial distribution.}
\usage{
range_probability(t, n, m)
}
\arguments{
\item{t}{A length-one numeric vector indicating the value to compute the
cumulative distribution function for.}

\item{n}{A length-one integer vector indicating the number of independent
balls.}

\item{m}{A length-one integer vector indicating the number of independent
urns/cells.}
}
\value{
A length-one numeric vector representing the probability of the
smallest order statistic.
}
\description{
This function calculates the cumulative distribution function (CDF)
of the range under the assumption of an equiprobable multinomial
distribution.
}
\examples{
range_probability(1, 7, 10)

}
\references{
Bonetti, M., Cirillo, P., Ogay, A. (2019), "Computing the exact
distributions of some functions of the ordered multinomial counts:
maximum, minimum, range and sums of order statistics", Royal Society
Open Science, 6: 190198, \url{http://dx.doi.org/10.1098/rsos.190198}.
}
\seealso{
\code{\link{highest_order_statistics}} for computing the
CDF of the sum of the first \eqn{J} largest order statistics.

\code{\link{smallest_order_value}} for computing the CDF
of the smallest order statistic.

\code{\link{highest_order_statistics}} for computing the sum of
the first \eqn{J} largest order statistics.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
