% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{XOMultinom-package}
\alias{XOMultinom-package}
\alias{XOMultinom-pkg}
\alias{XOMultinom-p}
\alias{XOMultinom}
\title{XOMultinom: a package}
\description{
The \pkg{XOMultinom} package contains functions for computing the exact
distributions of some functions of the ordered multinomial counts.

The main functions included in the package are:
\itemize{
\item \code{\link{highest_order_statistics}()}
\item \code{\link{max_order_statistic}()}
\item \code{\link{range_probability}()}
\item \code{\link{smallest_order_value}()}
}
}
\keyword{internal}
